+++ source/virtualBuffers/gecko_ia2.py
@@ -172,10 +172,10 @@ class Gecko_ia2_TextInfo(VirtualBufferTextInfo):
 			if controlTypes.State.LINKED not in states:
 				# This is a named link destination, not a link which can be activated. The user doesn't care about these.
 				role = controlTypes.Role.TEXTFRAME
-			elif (value := attrs.get("IAccessible::value")) is not None and self.obj.isInternalLinkInDocument(
-				value,
-			):
-				states.add(controlTypes.State.INTERNAL_LINK)
+			elif (value := attrs.get("IAccessible::value")) is not None and (
+				linkType := self.obj.getLinkTypeInDocument(value)
+			) is not None:
+				states.add(linkType)
 		level = attrs.get("IAccessible2::attribute_level", "")
 		xmlRoles = attrs.get("IAccessible2::attribute_xml-roles", "").split(" ")
 		landmark = next((xr for xr in xmlRoles if xr in aria.landmarkRoles), None)