+++ user_docs/en/changes.md
@@ -59,11 +59,11 @@ Please refer to [the developer guide](https://www.nvaccess.org/files/nvda/docume
 * It is now possible to redirect objects retrieved from on-screen coordinates, by using the `NVDAObject.objectFromPointRedirect` method. (#16788, @Emil-18)
 * Running SCons with the parameter `--all-cores` will automatically pick the maximum number of available CPU cores. (#16943, #16868, @LeonarddeR)
 * Developer info now includes information on app architecture (such as AMD64) for the navigator object. (#16488, @josephsl)
-* A new utils/urlUtils/isSamePageUrl function has been added to check links (#141, @LeonarddeR)
+* A new utils/urlUtils with different functions to determine link types has been added to check links (#141, @LeonarddeR, @nvdaes)
 * A new INTERNAL_LINK state has been added to controlTypes.states.State (#141, @nvdaes)
-* A new isInternalLink property has been added on `NVDAObject`. It queries the treeInterceptor by default, if any. (#141, @LeonarddeR, @nvdaes)
+* A new linkType property has been added on `NVDAObject`. It queries the treeInterceptor by default, if any. (#141, @LeonarddeR, @nvdaes)
 * BrowseModeTreeInterceptor object has a new documentUrl property (#141, @LeonarddeR)
-* BrowseModeTreeInterceptor object has a new isInternalLinkInDocument method which accepts an URL, to check if it corresponds to an internal link (#141, @LeonarddeR)
+* BrowseModeTreeInterceptor object has a new getLinkTypeInDocument method which accepts an URL to check the link type of the object (#141, @LeonarddeR, @nvdaes)
 
 
 #### Deprecations